migration: fix differences from wxt migration
