+++ extension/src/entrypoints/page.content.ts
@@ -1,4 +1,5 @@
 import { currentPageDelegate } from '@/services/pages';
+import { getAnimeTitleAndEpisode } from '@/services/anime-sites';
 import type { ContentScriptContext } from '#imports';
 
 const excludeGlobs = ['*://killergerbah.github.io/asbplayer*', '*://app.asbplayer.dev/*'];
@@ -16,5 +17,14 @@ export default defineContentScript({
 
     main(ctx: ContentScriptContext) {
         currentPageDelegate().then((pageDelegate) => pageDelegate?.loadScripts());
+
+        browser.runtime.onMessage.addListener((message, sender, sendResponse) => {
+            if (message.action === 'getTitleAndEp') {
+                getAnimeTitleAndEpisode(message.url)
+                    .then((result) => sendResponse(result))
+                    .catch((error) => sendResponse({ error: error.message }));
+                return true;
+            }
+        });
     },
 });
