--- extension/src/manifest.json
@@ -1,151 +1,0 @@
-{
-    "name": "API subs for asbplayer",
-    "description": "API subs for asbplayer",
-    "version": "1.0.1",
-    "manifest_version": 3,
-
-    "key": "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAxmdAa3ymqAjLms43ympXqtyuJnC2bSYh70+5ZZmtyx/MsnGhTEdfbqtsp3BKxHbv0rPd49+Joacm1Shik5/mCppZ0h4I4ISMm983X01H6p/hfAzQYAcnvw/ZQNHAv1QgY9JiuyTBirCDoYB50Fxol/kI/0EviYXuX83KoYpjB0VGP/ssY9ocT//fQUbRmeLDJnciry8y6MduWXHzseOP99axQIjeVsNTE30L4fRN+ppX3aOkG/RFJNx0eI02qbLul3qw5dUuBK5GgMbYftwjHnDoOegnZYFr1sxRO1zsgmxdp/6du75RiDPRJOkPCz2GTrw4CX2FCywbDZlqaIpwqQIDAQAB",
-    "default_locale": "en",
-    "icons": {
-        "16": "assets/icon16.png",
-        "48": "assets/icon48.png",
-        "128": "assets/icon128.png"
-    },
-    "minimum_chrome_version": "116",
-    "permissions": [
-        "tabs",
-        "tabCapture",
-        "activeTab",
-        "storage",
-        "contextMenus",
-        "sidePanel",
-        "offscreen",
-        "webNavigation",
-        "scripting"
-    ],
-    "host_permissions": ["<all_urls>"],
-    "action": {
-        "default_title": "asbplayer"
-    },
-    "options_ui": {
-        "page": "settings-ui.html",
-        "open_in_tab": true
-    },
-    "background": {
-        "service_worker": "background.js",
-        "type": "module"
-    },
-    "side_panel": {
-        "default_path": "side-panel.html"
-    },
-    "content_scripts": [
-        {
-            "matches": ["<all_urls>"],
-            "exclude_globs": ["*://killergerbah.github.io/asbplayer*", "*://localhost:3000/*"],
-            "js": ["video.js"],
-            "css": ["video.css"],
-            "all_frames": true,
-            "run_at": "document_idle"
-        },
-        {
-            "matches": ["<all_urls>"],
-            "exclude_globs": ["*://killergerbah.github.io/asbplayer*", "*://localhost:3000/*"],
-            "js": ["page.js"],
-            "all_frames": true,
-            "run_at": "document_start"
-        },
-        {
-            "matches": ["*://killergerbah.github.io/asbplayer*", "*://localhost/*"],
-            "js": ["asbplayer.js"],
-            "all_frames": true,
-            "run_at": "document_start"
-        }
-    ],
-    "commands": {
-        "copy-subtitle": {
-            "description": "__MSG_shortcutMineSubtitleDescription__"
-        },
-        "copy-subtitle-with-dialog": {
-            "suggested_key": {
-                "default": "Ctrl+Shift+X",
-                "mac": "MacCtrl+Shift+X"
-            },
-            "description": "__MSG_shortcutMineSubtitleAndOpenDialogDescription__"
-        },
-        "update-last-card": {
-            "suggested_key": {
-                "default": "Ctrl+Shift+U",
-                "mac": "MacCtrl+Shift+U"
-            },
-            "description": "__MSG_shortcutUpdateLastCardDescription__"
-        },
-        "toggle-video-select": {
-            "suggested_key": {
-                "default": "Ctrl+Shift+F",
-                "mac": "MacCtrl+Shift+F"
-            },
-            "description": "__MSG_shortcutSelectSubtitleTrackDescription__"
-        },
-        "take-screenshot": {
-            "suggested_key": {
-                "default": "Ctrl+Shift+V",
-                "mac": "MacCtrl+Shift+V"
-            },
-            "description": "__MSG_shortcutTakeScreenshotDescription__"
-        },
-        "toggle-recording": {
-            "description": "__MSG_shortcutToggleRecordingDescription__"
-        }
-    },
-    "web_accessible_resources": [
-        {
-            "resources": [
-                "chunks/*",
-                "assets/image.png",
-                "assets/fonts.css",
-                "assets/Roboto-Bold.woff2",
-                "assets/Roboto-Light.woff2",
-                "assets/Roboto-Medium.woff2",
-                "assets/Roboto-Regular.woff2",
-                "assets/Roboto-BoldItalic.woff2",
-                "assets/Roboto-LightItalic.woff2",
-                "assets/Roboto-MediumItalic.woff2",
-                "assets/Roboto-Italic.woff2",
-                "pages/netflix-page.js",
-                "pages/youtube-page.js",
-                "pages/anime-page.js",
-                "pages/tver-page.js",
-                "pages/bandai-channel-page.js",
-                "pages/amazon-prime-page.js",
-                "pages/hulu-page.js",
-                "pages/disney-plus-page.js",
-                "pages/apps-disney-plus-page.js",
-                "pages/viki-page.js",
-                "pages/unext-page.js",
-                "pages/emby-jellyfin-page.js",
-                "pages/osnplus-page.js",
-                "pages/bilibili-page.js",
-                "pages/nrk-tv-page.js",
-                "pages/plex-page.js",
-                "pages/areena-yle-page.js",
-                "anki-ui.js",
-                "mp3-encoder-worker.js",
-                "pgs-parser-worker.js",
-                "video-data-sync-ui.js",
-                "video-select-ui.js",
-                "mobile-video-overlay-ui.html",
-                "mobile-video-overlay-ui.js",
-                "notification-ui.js",
-                "asbplayer-locales/de.json",
-                "asbplayer-locales/en.json",
-                "asbplayer-locales/es.json",
-                "asbplayer-locales/ja.json",
-                "asbplayer-locales/pl.json",
-                "asbplayer-locales/zh_CN.json",
-                "asbplayer-locales/pt_BR.json",
-                "asbplayer-locales/ru.json"
-            ],
-            "matches": ["<all_urls>"]
-        }
-    ]
-}
